% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bas_hazard.R
\name{bas_hazard_internal}
\alias{bas_hazard_internal}
\title{Internal Function for the Baseline Hazard Step-Function}
\usage{
bas_hazard_internal(optimal_params, time_axis)
}
\arguments{
\item{optimal_params}{Numerical vector of length equal to the number of model parameters, containing the optimal estimated parameters.}

\item{time_axis}{Numerical vector of temporal domain.}
}
\value{
Numerical vector of length equal to the number of intervals of the time-domain, with the value of the baseline hazard step-function.
}
\description{
The method computes the baseline hazard step-function in each interval of the time-domain, using the estimated parameters
\eqn{\phi_k, \forall k}
}
\keyword{internal}
