% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{check.C_mult}
\alias{check.C_mult}
\title{Check positiveness of the multiplicative constant C}
\usage{
check.C_mult(C_mult)
}
\arguments{
\item{C_mult}{Multiplicative constant}
}
\value{
An error if the condition is not satisfied.
}
\description{
The method controls the multiplicative constant C is non-negative (i.e. positive).
}
