% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{check.centre}
\alias{check.centre}
\title{Check Correctness for the Cluster Variable}
\usage{
check.centre(centre)
}
\arguments{
\item{centre}{Numerical vector of length equal to the number of individuals in the study, containing the individual grouo/cluster
membership.}
}
\value{
An error if any condition is not satisfied.
}
\description{
The function controls that the provided cluster variable is a vector, with at least two levels. Indeed, it is not possible to
apply the Time-Dependent Shared Frailty Cox Model with no clusters.
}
\keyword{internal}
