% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{check.dataset}
\alias{check.dataset}
\title{Check Presence of NULL or NaN Element Value in the Dataset}
\usage{
check.dataset(data)
}
\arguments{
\item{data}{Dataset (dataframe)}
}
\value{
An error if any condition is not satisfied.
}
\description{
The method controls that the dataset does not contain 'NULL', 'null', 'NaN' or 'nan' value.
}
\keyword{internal}
