% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{check.flag_optimal_params}
\alias{check.flag_optimal_params}
\title{Check Coherence Between Flag for Optimal Parameters and Optimal Parameters}
\usage{
check.flag_optimal_params(optimal_params, flag_optimal_params)
}
\arguments{
\item{optimal_params}{Either a numerical vector of length equal to the number of model parameters or a NULL value.}

\item{flag_optimal_params}{Logical value. Did the user want to provide optimal parameters vector? If so, the variable should be TRUE; otherwise
(no optimal parameters), FALSE.}
}
\value{
An error if any condition is not satisfied.
}
\description{
The function controls that one of the following condition is satisfied:
\itemize{
\item if the flag for the optimal parameters is activated, then the optimal parameters should be provided in input
\item if the flag is not activated, then the optimal parameters should not be provided and the parameter vector should be NU
}
}
\keyword{internal}
