% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{check.frailty_dispersion}
\alias{check.frailty_dispersion}
\title{Check correctness of frailty standard deviation}
\usage{
check.frailty_dispersion(frailty_dispersion, n_intervals)
}
\arguments{
\item{sd}{Numerical vector of length equal to the number of intervals of the time-domain and containing the frailty standard deviation.}

\item{time_axis}{Numerical vector of the temporal domain subdivision.}
}
\value{
An error if any condition is not satisfied.
}
\description{
The function controls that the frailty standard deviation vector has a length equal to the number of inyervals of the time domain and
that its elements are non-negative.
}
