% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{check.index}
\alias{check.index}
\title{Check Existence of Provided Input Index}
\usage{
check.index(index, n_params)
}
\arguments{
\item{index}{Index with respect to which the user wants to study the one dimensional behaviour of
the log-likelihood function.}

\item{n_params}{Number of parameters of the model}
}
\value{
An error if any condition is not satisfied.
}
\description{
The method controls that the provided input index exists: it cannot be greater than the maximum number
of parameters of the current model.
}
\keyword{internal}
