% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{check.pchtype_colorbg}
\alias{check.pchtype_colorbg}
\title{Check Correctness of Plot Variables Pch and Color}
\usage{
check.pchtype_colorbg(centre_codes, pch_type, color_bg)
}
\arguments{
\item{centre_codes}{Numerical vector of length equal to the number of centres/clusters in the dataset and containing the distinct centres/clusters.
They correspond to the levels of the numerical vector of individual group membership.}

\item{pch_type}{Numerical vector of length equal to the number of centres and containing the point shape for each faculty.}

\item{color_bg}{Numerical vector of length equal to the number of centres and containing the color of the point for each faculty.}
}
\value{
An error if any condition is not satisfied.
}
\description{
The function controls that the input variables 'pch_type' and 'color_bg' have the correct structure,
they have the same dimension of the number of clusters in the dataset and they have meaningful elements.

These variables are used for the plot of the posterio frailty estimates: the estimates for each faculty are plotted through a symbol, having
color and shape indicated by the variables (for the k-th faculty, consider the k-th element of both vectors).
}
\keyword{internal}
