% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{check.pos_frailty_sd}
\alias{check.pos_frailty_sd}
\title{Check Positivity of the Frailty Standard Deviation}
\usage{
check.pos_frailty_sd(sd, n_intervals)
}
\arguments{
\item{sd}{Numerical vector of length equal to the number of intervals, containing the frailty standard deviation}

\item{n_intervals}{Number of intervals of the time-domain}
}
\value{
An error if any condition is not satisfied.
}
\description{
The method controls that the frailty standard deviation vector has non-negative elements
}
\keyword{internal}
