% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{check.post_frailty_centre}
\alias{check.post_frailty_centre}
\title{Check numerosity of posterior frailty estimates}
\usage{
check.post_frailty_centre(post_frailty_est, centre_codes)
}
\arguments{
\item{post_frailty_est}{An S3 class object containing the posterior frailty estimates \eqn{\hat{\alpha}_j, \hat{\epsilon}_{jk},
\hat{Z}_{jk}, \forall j,k}}

\item{centre_codes}{Numerical vector of length equal to the number of distinct centres/clusters in the study}
}
\value{
An error if any condition is not satisfied.
}
\description{
The function controls that a time-dependent posterior frailty estimate is computed for each centre
}
