% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{check.structure_paramsCI}
\alias{check.structure_paramsCI}
\title{Check structure for the Parameters Confidence Interval}
\usage{
check.structure_paramsCI(parametersCI)
}
\arguments{
\item{parametersCI}{S3 object of class 'ParametersCI', composed of two elements:
\itemize{
\item left confidence interval: numerical vector of length equal to the number of parameters in the model
\item right confidence interval: numerical vector of length equal to the number of parameters in the model
}}
}
\value{
An error if any condition is not satisfied.
}
\description{
The function controls that the structure of the Parameters Confidence Intervals coincides with the theoretical one.
}
