% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{check.structure_post_frailty_CI}
\alias{check.structure_post_frailty_CI}
\title{Check structure of Posterior Frailty Confidence Interval}
\usage{
check.structure_post_frailty_CI(post_frailty_CI, n_intervals, n_centres)
}
\arguments{
\item{post_frailty_CI}{Posterior frailty estimates S3 object of class 'PFCI.AdPaik', composed of two elements:
\itemize{
\item left confidence interval for the estimated \eqn{\hat{Z}_{jk}, \forall j,k}
\item right confidence interval for the estimated \eqn{\hat{Z}_{jk}, \forall j,k}
}}

\item{n_intervals}{Number of intervals of the time-domain}

\item{n_centres}{Number of centres/clusters.}
}
\value{
An error if any condition is not satisfied.
}
\description{
The function controls that the structure of the 'Posterior Frailty Confidence Interval' coincides with the theoretical one.
}
