% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{check.time_axis}
\alias{check.time_axis}
\title{Check correctness of time domain subdivision}
\usage{
check.time_axis(time_axis)
}
\arguments{
\item{time_axis}{Numerical vector of temporal domain subdivision.}
}
\value{
An error is returned if any condition is not satisfied.
}
\description{
The function controls that the time domain is a vector and it has at least 2 elements and that all of them are not negative.
Moreover, it checks that all its elements are non-negative and properly ordered, in an ascending way.
}
