% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{check.value_post_frailty}
\alias{check.value_post_frailty}
\title{Check non-negativeness of the posterior frailty estimates}
\usage{
check.value_post_frailty(post_frailty_est, n_centres, n_intervals)
}
\arguments{
\item{post_frailty_est}{An S3 class object containing the posterior frailty estimates: \eqn{\hat{\alpha}_j, \hat{\epsilon}_{jk},
\hat{Z}_{jk}, \forall j,k}}

\item{n_centres}{Number of groups/clusters.}

\item{n_intervals}{Number of intervals of the time domain}
}
\value{
An error if any condition is not satisfied.
}
\description{
The function controls that all posterior frailty estimates are non-negative.
Indeed, by construction the realizations of a gamma distribution are non negative.
}
