% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractors.AdPaik.R
\name{extractAIC.AdPaik}
\alias{extractAIC.AdPaik}
\title{Extract AIC for \code{AdPaik} Objects}
\usage{
\method{extractAIC}{AdPaik}(fit, scale = NULL, k = 2, ...)
}
\arguments{
\item{fit}{An \code{AdPaik} model object.}

\item{scale}{Changing it is not supported for this model. It will be ignored.}

\item{k}{Penalty parameter (default is 2 for AIC).}

\item{...}{Additional arguments (ignored).}
}
\value{
A numeric vector with the number of parameters and AIC value.
}
\description{
Computes the AIC for an \code{AdPaik} model.
}
