% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractors.AdPaik.R
\name{formula.AdPaik}
\alias{formula.AdPaik}
\title{Extract Model Formula for \code{AdPaik}}
\usage{
\method{formula}{AdPaik}(x, ...)
}
\arguments{
\item{x}{An \code{AdPaik} model object.}

\item{...}{Additional arguments (ignored).}
}
\value{
The formula object.
}
\description{
Retrieves the formula used in the \code{AdPaik} model.
}
