% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frailty_mean.R
\name{frailty_mean.PowPar}
\alias{frailty_mean.PowPar}
\title{Estimated frailty mean for the 'Centre-Specific Frailty Model with Power Parameter'}
\usage{
frailty_mean.PowPar(optimal_params, time_axis, n_regressors)
}
\arguments{
\item{optimal_params}{Numerical vector of optimal estimated parameters, of length equal to the number of
model parameters.}

\item{time_axis}{Numerical vector of the time-domain.}

\item{n_regressors}{Number of regressors of the model}
}
\value{
Numerical vector of length equal to the number of intervals, containing the estimated frailty
mean of each interval.
}
\description{
The function computes the model estimated frailty mean, giving the estimated parameters.
}
