% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frailty_sd.R
\name{frailty_sd}
\alias{frailty_sd}
\title{Function for computing frailty standard deviation and Variance}
\usage{
frailty_sd(result, flag_fullsd = TRUE)
}
\description{
According to the model it has been called,
this function inspects the class output and then call the proper summary function.
}
