% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractors.AdPaik.R
\name{getData.AdPaik}
\alias{getData.AdPaik}
\title{Extract Processed Dataset from an AdPaik Object}
\usage{
getData.AdPaik(object)
}
\arguments{
\item{object}{An \code{AdPaik} model object.}
}
\value{
The processed dataset used in the model, or \code{NULL} if no dataset is found.
}
\description{
Retrieves the dataset used in the 'Adapted Paik et al.' model, applying any
missing value handling specified in the model object.
}
