% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AdPaikModel.R
\name{ll_AdPaik_centre_eval}
\alias{ll_AdPaik_centre_eval}
\title{Evaluation of Model Group log-Likelihood}
\usage{
ll_AdPaik_centre_eval(params, dataset, dropout_matrix, e_matrix)
}
\arguments{
\item{params}{Parameter vector.}

\item{dataset}{Matrix of dataset regressors, with a number of rows equal to the number of individuals in a cluster.}

\item{dropout_matrix}{Binary matrix indicating in which interval of the time domain and individual failed. For an individual,
the sum of the row elements must be equal to 1 (if he/she failed) or 0 (if he/she does not failed).
It has dimension equal to (n_individuals, n_intervals)}

\item{e_matrix}{Matrix of dimension (n_individual, n_intervals), where each element contains the evaluation of the temporal
integral, performed through the function @time_int_eval.}
}
\value{
Group log-likelihood evaluation
}
\description{
Evaluation of model group log-likelihood at the provided parameter vector and data.
This function is internally called by 'll_AdPaik_eval' to evaluate the log-likelihood function, considering
all and only the individuals belonging to a group.
}
\keyword{internal}
