% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractors.AdPaik.R
\name{logLik.AdPaik}
\alias{logLik.AdPaik}
\title{Extract Log-Likelihood for \code{AdPaik} Objects}
\usage{
\method{logLik}{AdPaik}(object, ...)
}
\arguments{
\item{object}{An \code{AdPaik} model object.}

\item{...}{Additional arguments (ignored).}
}
\value{
A log-likelihood object with degrees of freedom (\code{df}).
}
\description{
Returns the log-likelihood of the fitted \code{AdPaik} model.
}
