% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NodesWeightsQuadratureFomula.R
\docType{data}
\name{n_nodes}
\alias{n_nodes}
\title{Nodes and weights for the Gauss_hermite quadrature formula
for the 'Centre-Specific Frailty Model with Power Parameter'.
The nodes and weights have been extracted from the 'Handbook of Mathematical functions'
pag 940.}
\format{
An object of class \code{numeric} of length 1.
}
\usage{
n_nodes
}
\description{
Nodes and weights for the Gauss_hermite quadrature formula
for the 'Centre-Specific Frailty Model with Power Parameter'.
The nodes and weights have been extracted from the 'Handbook of Mathematical functions'
pag 940.
}
\keyword{datasets}
