% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NodesWeightsQuadratureFomula.R
\docType{data}
\name{n_nodesG}
\alias{n_nodesG}
\title{Nodes and weights for the Gauss-Hermite quadrature formula,
for the 'Stochastic Time-Dependent Centre-Specific Frailty Model'.
For the G function, the chosen nodes should not contain the zero (node)
since it appears at the denominator of a fraction.
Also in this case, the nodes and weights have been extracted from the
'Handbook of Mathematical functions', pag 940.}
\format{
An object of class \code{numeric} of length 1.
}
\usage{
n_nodesG
}
\description{
Nodes and weights for the Gauss-Hermite quadrature formula,
for the 'Stochastic Time-Dependent Centre-Specific Frailty Model'.
For the G function, the chosen nodes should not contain the zero (node)
since it appears at the denominator of a fraction.
Also in this case, the nodes and weights have been extracted from the
'Handbook of Mathematical functions', pag 940.
}
\keyword{datasets}
