% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractors.AdPaik.R
\name{nobs.AdPaik}
\alias{nobs.AdPaik}
\title{Extract Number of Observations for \code{AdPaik}}
\usage{
\method{nobs}{AdPaik}(object, ...)
}
\arguments{
\item{object}{An \code{AdPaik} model object.}

\item{...}{Additional arguments (ignored).}
}
\value{
Integer: Number of observations.
}
\description{
Returns the number of observations used in the model.
}
