% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.ll_1D.AdPaik}
\alias{plot.ll_1D.AdPaik}
\title{Plot the one-dimensional log-likelihood function}
\usage{
\method{plot}{ll_1D.AdPaik}(
  param_1D,
  index_param_1D,
  ll_1D,
  params,
  param_range_min,
  param_range_max,
  dataset,
  centre,
  time_axis,
  dropout_matrix,
  e_matrix,
  n_points = 150,
  cex = 0.7,
  cex_max = 0.8,
  color_bg = "black",
  color_max_bg = "red",
  pch = 21
)
}
\arguments{
\item{param_1D}{Optimal parameter value determined maximizing the log-likelihood function with respect to it.}

\item{index_param_1D}{Index of the optimal parameter inside the parameter vector.}

\item{ll_1D}{Log-likelihood function evaluated at the optimal parameter @param_1D, with the other parameters assuming a fixed value.}

\item{params}{Numerical vector of length equal to the number of parameters minus one. It contains the fixed value of the other parameters.}

\item{param_range_min}{Minimum value assumable by the parameter param_1D.}

\item{param_range_max}{Maximum value assumable by the parameter param_1D.}

\item{dataset}{Dataset with individual covariates. It can be either a matrix or a dataframe.}

\item{centre}{Numerical vector of individual cluster membership. It must be length as the number of individual in the dataset.}

\item{time_axis}{Numerical vector of length 2, corresponding to the subdivision of the temporal domain.}

\item{dropout_matrix}{Binary matrix indicating in which interval of the time domain an individual failed. For each individual,
the sum of the row elements must be equal to 1 (if he/she failed) or 0 (if he/she does not failed).
It has dimension equal to (n_individuals, n_intervals)}

\item{e_matrix}{Matrix of dimension (n_individual, n_intervals), where each element contains the evaluation of the temporal
integral, performed through the function @time_int_eval.}

\item{n_points}{Number of points in which the log-likelihood function must be evaluated and then plotted.
To have a nice graphical representation, chose an intermediate value: not to small and not too high. Default value is 150.}

\item{cex}{Dimension of the points used for the graphical representation of the log-likelihood function. Default value is 0.7.}

\item{cex_max}{Dimension of the optimal point provided as first argument (i.e. the point that maximizes the log-likelihood
function from the point of view of a single parameter). Default value is 0.8.}

\item{color_bg}{Color to be used for the points of the log-likelihood trend. Deafult is 'black'.}

\item{color_max_bg}{Color to be used for the optimal point provided as first argument. Default is 'red'.}

\item{pch}{Shape of the plotted point. Deafult is a circle (21).}
}
\value{
Plot of the log-likelihood trend from the point of view of a single parameter.
}
\description{
Function for plotting the trend of the log-likelihood function with respect to
a single parameter, whose position in the parameter vector is specified in the argument call.
For such plot, some samples of the parameter under consideration are generated and
then evaluated in the log-likelihood function. These are plotted together with the maximum point of the
one-dimensional log-likelihood function, provided in input.
}
