% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.ll_1D.StocTimeDep}
\alias{plot.ll_1D.StocTimeDep}
\title{Function for plotting the trend of the log-likelihood function with respect to
a single parameter, specified in the arguments.
The log-likelihood function is plotted using a certain number of points specified by @n_points
and they are not connected by a straight line.}
\usage{
\method{plot}{ll_1D.StocTimeDep}(
  param_1D,
  index_param_1D,
  ll_1D,
  params,
  param_range_min,
  param_range_max,
  dataset,
  centre,
  time_axis,
  dropout_matrix,
  e_matrix,
  time_to_event,
  n_points = 150,
  cex = 0.7,
  cex_max = 0.8,
  color_bg = "black",
  color_max_bg = "red",
  pch = 21
)
}
\arguments{
\item{_1D}{Optimal parameter value determined maximizing the log-likelihood function with respect to it.}

\item{_range_min}{Minimum value assumable by the parameter param_1D.}

\item{_range_max}{Maximum value assumable by the parameter param_1D.}
}
\description{
Function for plotting the trend of the log-likelihood function with respect to
a single parameter, specified in the arguments.
The log-likelihood function is plotted using a certain number of points specified by @n_points
and they are not connected by a straight line.
}
