% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_frailty_sd}
\alias{plot_frailty_sd}
\title{Plot for the frailty standard deviation}
\usage{
plot_frailty_sd(
  result,
  frailty_sd = NULL,
  flag_variance = FALSE,
  flag_sd_external = FALSE,
  xlim = c(1, length(time_axis) - 1),
  ylim = c(0, 10),
  xlab = "Intervals",
  ylab = "Values",
  main_title = "Frailty standard deviation",
  pch = 21,
  color_bg = "blue",
  cex_points = 0.7
)
}
\arguments{
\item{result}{S3 object of class 'AdPaik', returned by the main model call 'AdPaikModel(...)'.}

\item{frailty_sd}{Numerical vector of evaluated frailty standard deviation, of length equal to the number of intervals of the time-domain.
Its element must be non-negative.}

\item{flag_variance}{Do you want to plot the frailty standard deviation? If so, the flag must be equal to TRUE; otherwise, to FALSE (if we want to
plot the variance). Default value is FALSE.}

\item{flag_sd_external}{Logical value. Do you want to provide a frailty standard deviation}

\item{xlim}{Numerical vector of length 2, giving the x coordinate ranges.
Default value goes from 1 up to the number of intervals of the time-domain, because for each interval a single point is plotted and not a segment.}

\item{ylim}{Numerical vector of length 2, giving the y coordinate ranges.
Default value is (0,10).}

\item{xlab, ylab}{String giving the x and y axis name.
Deafult values are 'Intervals' and 'Values'.}

\item{main_title}{Title of the plot. Default title is 'Frailty standard deviation'.}

\item{pch}{Symbol used for plotting the frailty standard deviation value. Default is a dot (21).
The same symbol is used throughout the plot, with no possibility of change.}

\item{color_bg}{Color used for the symbols. Deafult is 'blue'.}

\item{cex_points}{Dimension of the symbols. Deafult is 0.7.}
}
\value{
Plot of either the frailty standard deviation or the frailty variance.
}
\description{
Function for plotting either the frailty standard deviation or the frailty variance, according to the value
of a precise and user-provided flag.
}
\details{
For each interval of the time-domain, a single point is plotted with value at the computed frailty standard deviation or variance.
Subsequent points are connected to show the trend of the required function.

The method gives the possibility of plotting either
\itemize{
\item the frailty standard deviation returned by the main model call and, therefore, contained in the S3 object class 'AdPaik'
or
\item the frailty standard deviation computed through the method 'frailty.sd'. We recall this method is introduced to permit the user to change the
way the frailty standard deviation has been computed, without the necessity of optimizing again the log-likelihood function.
For instance, we require the computation of the frailty standard deviation with both terms (time-dependent and time-independent) in the main model
'AdPaikModel', but now we want also the frailty standard deviation with the sole time-dependent term. Thanks to the introduced method 'frailty.sd',
we can compute again the frailty standard deviation without maximizing again the log-likelihood function.
The result we obtain in the second case is different in values and structure, because we simply get a vector of L elements instead
of a class, and we want the plot to be able to distinguish the case and plot it anyway.
}
}
\examples{
# Define the variables needed for the model execution
formula <-
time_axis <-
categories_range_min <-
categories_range_max <-

# Call the main model
result <- AdPaikModel(formula, data, time_axis,
                      categories_range_min, categories_range_max, TRUE)


plot_frailty_sd(sd, time_axis, FALSE, ylim = c(0,1))
}
