% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_post_frailty_est}
\alias{plot_post_frailty_est}
\title{Plot the posterior frailty estimates}
\usage{
plot_post_frailty_est(
  result,
  data,
  flag_eps = FALSE,
  flag_alpha = FALSE,
  xlim = c(1, length(time_axis) - 1),
  ylim = c(0, 10),
  xlab = "Intervals",
  ylab = "Values",
  main_title = "Posterior frailty estimates",
  cex = 0.7,
  pch_type = rep(21, length(centre_codes)),
  color_bg = rep("black", length(centre_codes)),
  cex_legend = 0.7,
  pos_legend = "topright"
)
}
\arguments{
\item{result}{S3 object of class 'AdPaik', returned by the method call 'AdPaikModel(...)'.}

\item{data}{Dataset (dataframe) in which all variables of the formula object must be found and contained.}

\item{flag_eps}{Does the user want to plot only the time-dependent posterior frailty estimates? If so, the flag must be TRUE.
Otherwise, for the entire posterior estimates, it must be FALSE (default value).}

\item{flag_alpha}{Does the user want to plot only the time-independent posteior frailty estimates? If so, the flag must be TRUE.
Otherwise, for the entire posterior estimates, it must be FALSE (default value).
Note that, it is not possible to have both previous flags TRUE: either one of the two must be TRUE. However, both can be FALSE.}

\item{xlim}{Numeric vector of length 2, giving the x coordinate ranges.
Default value goes from 1 up to the number of intervals of the time-domain, because a point is plotted for each faculty in each interval.}

\item{ylim}{Numeric vector of length 2, giving the y coordinate ranges.
Default value is (0,10). Pay attention that no non-negative values only can be accepted.}

\item{xlab, ylab}{String giving the x and y axis name. Default values are 'Intervals' and 'Values'.}

\item{main_title}{Title of the plot. Default title is 'Posterior frailty estimates'.}

\item{cex}{Dimension of the points used for plotting the estimates.}

\item{pch_type}{Numerical vector of length equal to the number of clusters in the data, giving the symbol to be used for plotting the estimates.
Default symbol (circle, 21) is the same for all clusters.}

\item{color_bg}{Numerical vector of length equal to the number of clusters in the data, giving the color to be used for plotting the symbols
for the estimates. Default ('black') is the same for all faculties. On the other hand, the same color is used throughout the intervals for
the same faculty.}

\item{cex_legend}{Dimension of the symbol in the legend. Default is 0.7.}

\item{pos_legend}{It could be either a numerical vector of length 2, providing the x and y coordinates of the legend, or a string indicating where
the legend has to be located in the plot. In the last case, the only possibilities are: 'bottomright', 'bottom', 'bottomleft', 'left',
'topleft', 'top', 'topright', 'right', 'center'.}
}
\value{
The plot of the posterior frailty estimates.
}
\description{
Function for plotting the posterior frailty estimate of each group in each interval,
using a single point. For the same group, the sequence of points is then connected using a straight
line.
}
\details{
Recalling the frailty structure as \eqn{Z_{jk} = \alpha_j + \epsilon_{jk}, \forall j,k} and the posterior
frailty estimate as \eqn{\hat{Z}_{jk} = \hat{\alpha}_j/\hat{\alpha}_{max} + \hat{\epsilon}_{jk}/\hat{\epsilon}_{max}}, it is
possible to plot the entire posterior frailty estimate \eqn{\hat{Z}_{jk}} or just a part: either the time-dependent estimate
\hat{\epsilon}\emph{{jk}/\hat{\epsilon}}{max} or the time-independent one \hat{\alpha}\emph{j/\hat{\alpha}}{max}.
To do so, the user has to specify the alue of two flags.
}
\examples{
# Import data
data(data_dropout)

# Define the variables needed for the model execution
eps_paik <- 1e-10
categories_range_min <- c(-8, -2, eps_paik, eps_paik, eps_paik)
categories_range_max <- c(-eps_paik, 0.4, 1 - eps_paik, 1, 10)
time_axis <- c(1.0, 1.4, 1.8, 2.3, 3.1, 3.8, 4.3, 5.0, 5.5, 5.8, 6.0)
formula <- time_to_event ~ Gender + CFUP + cluster(group)

# Call the main model function
result <- AdPaikModel(formula, data, time_axis, categories_range_min, categories_range_max)

# Define variables for plotting the estimates
pch_type <- c(21, seq(21,25,1), seq(21,25,1), seq(21,25,1))
color_bg <- c("darkblue", rep("red", 5), rep("purple", 5), rep("green",5))

plot_post_frailty_est(result, data_dropout, ylim=c(0,2),
                      pch_type = pch_type, color_bg = color_bg)
}
