% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posterior_frailty.R
\name{post_frailty_CI_internal}
\alias{post_frailty_CI_internal}
\title{Confidence Interval for Posterior Frailty Estimates}
\usage{
post_frailty_CI_internal(
  post_frailty_est,
  post_frailty_est_var,
  n_centres,
  n_intervals,
  level
)
}
\arguments{
\item{post_frailty_est}{Posterior frailty estimates list.}

\item{post_frailty_est_var}{Posterior frailty variance list.}

\item{n_centres}{Number of clusters/centres.}

\item{n_intervals}{Number of intervals of the time-domain. it is equal to the length of the tima_axis minus one.}

\item{level}{A numeric value representing the confidence level.}
}
\value{
S3 object of class 'PFCI.AdPaik' composed of two matrices of dimension (number groups, number of intervals):
\itemize{
\item PostFrailtyCI_left: left confidence interval for each posterior frailty estimates
\item PostFrailtyCI_right: right confidence interval for each each posterior frailty estimates
}
}
\description{
Function for computing the confidence interval for each posterior frailty estimates \eqn{\hat{Z}_{jk}}.
}
\keyword{internal}
