% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posterior_frailty.R
\name{post_frailty_est}
\alias{post_frailty_est}
\title{Posterior Frailty Estimates}
\usage{
post_frailty_est(object, flag_eps = FALSE, flag_alpha = FALSE)
}
\arguments{
\item{object}{S3 object of class 'AdPaik' returned by the main model output, that contains all the information for the computation
of the frailty standard deviation.}

\item{flag_eps}{Logical flag indicating whether to plot only the time-dependent posterior frailty estimates. Default is FALSE.}

\item{flag_alpha}{Logical flag indicating whether to plot only the time-independent posterior frailty estimates. Default is FALSE.}
}
\value{
Vector or matrix of posterior frailty estimates, depending on the flag_eps and flag_alpha values.
Specifically:
\itemize{
\item 'alpha': posterior frailty estimates for \eqn{\alpha_j, \forall j}. It is a vector of length equal to the number of centres.
\item 'eps': posterior frailty estimates for \eqn{\epsilon_{jk}, \forall j,k}. Matrix of dimension (N, L).
\item 'Z': posterior frailty estimates for \eqn{Z_{jk} = \alpha_j + \epsilon_{jk}, \forall j,k}. Matrix of dimension (N, L).
}
}
\description{
Function for computing the posterior frailty estimates and variances of the time-dependent shared frailty Cox model.
Recalling the structure of the frailty \eqn{Z_{jk} = \alpha_j + \epsilon_{jk}, \forall j,k} as being composed by the sum
of two independent gamma distributions:
\itemize{
\item \eqn{\alpha_j \sim gamma(\mu_1/\nu, 1/\nu), \forall j}
\item \eqn{\epsilon_{jk} \sin gamma(\mu_2/\gamma_k, 1/\gamma_k), \forall j,k}
the posterior distribution of both terms is still a gamma with different mean and variance and the
posterior frailty estimate corresponds to the 'empirical Bayes estimate', that is the previous mentioned posterior mean.
}
}
\examples{
# Consider the 'Academic Dropout dataset'
data(data_dropout)

# Define the variables needed for the model execution
formula <- time_to_event ~ Gender + CFUP + cluster(group)
time_axis <- c(1.0, 1.4, 1.8, 2.3, 3.1, 3.8, 4.3, 5.0, 5.5, 5.8, 6.0)
eps <- 1e-10
categories_range_min <- c(-8, -2, eps, eps, eps)
categories_range_max <- c(-eps, 0, 1 - eps, 1, 10)

\donttest{
# Call the main model
result <- AdPaikModel(formula, data_dropout, time_axis,
                      categories_range_min, categories_range_max)

post_frailty_est(result)
}
}
