% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_and_print.R
\name{print.AdPaik}
\alias{print.AdPaik}
\title{Print method for AdPaik objects}
\usage{
\method{print}{AdPaik}(x, ...)
}
\arguments{
\item{x}{An object of class \code{AdPaik}.}

\item{...}{Additional arguments (not used).}
}
\value{
Prints a summary of the \code{AdPaik} object and returns it invisibly.
}
\description{
This function prints a summary of the optimal parameters estimated in an AdPaik object.
}
