% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{print.summary.AdPaik}
\alias{print.summary.AdPaik}
\title{Print Method for summary.AdPaik Objects}
\usage{
\method{print}{summary.AdPaik}(x, ...)
}
\arguments{
\item{x}{An object of class \code{"summary.AdPaik"}, created by \code{summary(object)}.}

\item{...}{Additional arguments (currently unused).}
}
\value{
\code{print.summary.AdPaik()} prints the formatted summary to the console.
}
\description{
\code{print} method for objects of class \code{"summary.AdPaik"}. Prints the formatted summary to the console.
}
\seealso{
\code{\link{summary}}, \code{\link{print}}
}
