% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survival.R
\name{survival}
\alias{survival}
\title{Compute Survival Function}
\usage{
survival(result, data)
}
\arguments{
\item{result}{S3 object of class 'AdPaik' containing model results.}

\item{data}{Data frame containing the dataset with covariates used in the model.}
}
\value{
A dataset where each row corresponds to the survival function values
over the time intervals for each individual in the dataset.
The first column represents the cluster to which the individual belongs.
}
\description{
Computes the survival function based on the 'Adapted Paik et al.' model's
given the estimated coefficients and frailty effects.
}
