% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_int_eval.R
\name{time_int_eval}
\alias{time_int_eval}
\title{Resolution of integral with respect to time}
\usage{
time_int_eval(time_t, k, time_axis)
}
\arguments{
\item{time_t}{Event time instant}

\item{k}{k-th interval of the time-axis}

\item{time_axis}{Temporal domain (it may coincide with the follow-up)}
}
\value{
Evaluation of the temporal integral
}
\description{
Function for the resolution of an integral with respect to time,
in the evaluation of the log-likelihood function.
It is implemented as defined in the paper of Wintrebert et al.'s (2004)
}
