% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TimeDepFrailty.R
\name{TimeDepFrail}
\alias{TimeDepFrail}
\title{PowParModel' = PowParModel(formula, data, time_axis,
StocTimeDepModel' = StocTimeDepModel(formula, data, time_axis,}
\usage{
TimeDepFrail(
  formula,
  data,
  time_axis,
  categories_range_min,
  categories_range_max,
  C_mult = 0,
  flag_fullsd = TRUE,
  time_domain = 0,
  flag_time_domain = FALSE,
  n_extrarun = 60,
  tol_ll = 1e-06,
  tol_optimize = 1e-06,
  h_dd = 0.001,
  print_previous_ll_values = c(TRUE, 3),
  model_type = c("AdPaikModel")
)
}
\description{
PowParModel' = PowParModel(formula, data, time_axis,
StocTimeDepModel' = StocTimeDepModel(formula, data, time_axis,
}
