% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{check.frailty_dispersion}
\alias{check.frailty_dispersion}
\title{Check Correctness of Frailty Standard Deviation}
\usage{
check.frailty_dispersion(frailty_dispersion, n_intervals)
}
\arguments{
\item{frailty_dispersion}{Frailty dispersion}

\item{n_intervals}{Number of intervals of the time-domain}
}
\value{
An error if any condition is not satisfied.
}
\description{
The function controls that the frailty standard deviation vector has a length equal to the number of inyervals of the time domain and
that its elements are non-negative.
}
\keyword{internal}
