% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frailty_sd.R
\name{frailty_sd}
\alias{frailty_sd}
\title{Frailty Standard Deviation and Variance}
\usage{
frailty_sd(object, ...)
}
\arguments{
\item{object}{An object for which frailty standard deviation and variance should be computed.}

\item{...}{Additional arguments to be passed to methods.}
}
\value{
An S3 class object 'FrailtyDispersion' containing:
\item{FrailtyVariance}{A numerical vector of frailty variances.}
\item{FrailtyStandardDeviation}{A numerical vector of frailty standard deviations.}
}
\description{
Computes both the standard deviation and the variance of the time-dependent
frailty for various models.
}
\seealso{
\code{\link{frailty_sd.AdPaik}}
}
