% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survival.R
\name{plot_survival}
\alias{plot_survival}
\title{Compute Survival Function}
\usage{
plot_survival(
  result,
  survival_df,
  lwd = 1,
  xlab = "Time",
  ylab = "Survival",
  main = "Stepwise Survival"
)
}
\arguments{
\item{result}{S3 object of class 'AdPaik' containing model results.}

\item{survival_df}{The dataframe returned by the survival function}
}
\value{
A dataset where each row corresponds to the survival function values
over the time intervals for each individual in the dataset.
}
\description{
Computes the survival function based on the 'Adapted Paik et al.' model's
estimated coefficients and frailty effects.
}
