% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_and_print.R
\name{print.summary.AdPaik}
\alias{print.summary.AdPaik}
\title{Print Method for summary.AdPaik Objects}
\usage{
\method{print}{summary.AdPaik}(x, ...)
}
\arguments{
\item{x}{An object of class \code{"summary.AdPaik"}.}

\item{...}{Additional arguments (currently unused).}
}
\value{
Prints a structured summary of the \code{AdPaik} model.
}
\description{
\code{print} method for objects of class \code{"summary.AdPaik"}. Formats and prints the model summary.
}
