% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary.AdPaik}
\alias{summary.AdPaik}
\title{Summary of the 'Adapted Paik et al.'s Model'}
\usage{
summary.AdPaik(result)
}
\arguments{
\item{result}{'S3' class object returned by the main model call, i.e. output of the 'Adapted Paik et al.'s Model'.}
}
\value{
Model summary printed on output.
}
\description{
Summary function for summarizing the most important information related to the dataset (number of individuals,
number of regressors, number of intervals, number of clusters), the model call (number of parameters) and the model
output (optimal log-likelihood value and AIC).
}
\details{
Among the estimated parameters, only the regressors are reported together with their standard error and confidence interval.
}
\examples{
# Define the variables needed for the model execution
formula <-
time_axis <-
categories_range_min <- c()
categories_range_max <- c()

# Call the main model function
result <- AdPaikModel(formula, data, time_axis, categories_range_min, categories_range_max, flag_fullsd = TRUE)

# Call the summary
summary.AdPaik(result)
}
