% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary.AdPaik}
\alias{summary.AdPaik}
\title{Summary of the Adapted Paik et al.'s Time-Dependent Shared Frailty Model}
\usage{
## S3 method for class 'AdPaik'
summary(object, ...)

## S3 method for class 'summary.AdPaik'
print(x, ...)
}
\arguments{
\item{object}{An object of class \code{"AdPaik"}, returned by the main model function.}

\item{x}{An object of class \code{"summary.AdPaik"}, created by \code{summary.AdPaik()}.}

\item{...}{Additional arguments (currently unused).}
}
\value{
\itemize{
\item \code{summary.AdPaik()} returns an object of class \code{"summary.AdPaik"}, containing structured model summary information.
\item \code{print.summary.AdPaik()} prints the formatted summary to the console.
}
}
\description{
Provides a structured summary of the results from the Adapted Paik et al.'s Time-Dependent Shared Frailty Model.
The summary includes key dataset information (e.g., number of individuals, regressors, intervals, and clusters),
model parameters, and main output metrics such as log-likelihood and AIC. Estimated regressors and their standard errors
are also reported.
}
\details{
This function extracts relevant model outputs, such as estimated parameters, standard errors,
and the convergence status of the algorithm. The summary is returned as an object of class \code{"summary.AdPaik"},
which can be printed using \code{print()}.
}
\examples{
\dontrun{
data(data_dropout)
eps_paik <- 1e-10
categories_range_min <- c(-8, -2, eps_paik, eps_paik, eps_paik)
categories_range_max <- c(-eps_paik, 0.4, 1 - eps_paik, 1, 10)
time_axis <- c(1.0, 1.4, 1.8, 2.3, 3.1, 3.8, 4.3, 5.0, 5.5, 5.8, 6.0)
formula <- time_to_event ~ Gender + CFUP + cluster(group)

# Fit the model
result <- AdPaikModel(formula, data_dropout, time_axis, categories_range_min, categories_range_max)

# Generate and print summary
summary_result <- summary(result)
print(summary_result)  # or simply `summary(result)`
}
}
