% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_and_print.R
\name{summary.AdPaik}
\alias{summary.AdPaik}
\title{Summary Method for AdPaik Objects}
\usage{
\method{summary}{AdPaik}(object, ...)
}
\arguments{
\item{object}{An object of class \code{"AdPaik"}, returned by the main model function.}

\item{...}{Additional arguments (currently unused).}
}
\value{
An object of class \code{"summary.AdPaik"} containing structured model summary information.
}
\description{
\code{summary} method for objects of class \code{"AdPaik"}. It prepares a structured summary of the results.
}
