% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary}
\alias{summary}
\title{Function for displaying on the R console a summary of the model call, given the
output of the same call.}
\usage{
summary(result)
}
\description{
According to the model it has been called, this function inspects the class output and then call
the proper summary function.
}
