% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary}
\alias{summary}
\title{Summary for Time-Dependent Frailty Models}
\usage{
summary(result)
}
\arguments{
\item{result}{An object containing the output of the model call.
The class of this object determines which summary method is invoked.}
}
\value{
A summary of the model output printed to the console.
}
\description{
This function displays a summary of the model output based on the class of the result object.
It delegates to the appropriate summary method according to the class of the result.
}
